Renaming the files.
