+++ src/app/grid/grid-select/grid-select-sample.component.html
@@ -1,29 +1,29 @@
 <div class="grid__wrapper">
     <div class="sample__header">
-        <igx-grid #grid1 [data]="data" [primaryKey]="'name'" width="100%" height="600px">
-            <igx-column field="name" header="Character name" [editable]="true" width="22%">
+        <igx-grid #grid1 [data]="data" [primaryKey]="'name'" width="780px" height="600px">
+            <igx-column field="name" header="Character name" [editable]="true" width="25%">
             </igx-column>
-            <igx-column field="race" header="Race" [editable]="true" [filterable]="false" width="22%">
+            <igx-column field="race" header="Race" [editable]="true" [filterable]="false" width="20%">
                 <ng-template igxCellEditor let-cell="cell" let-value>
-                    <igx-select [(ngModel)]="cell.editValue" width="100px" [igxFocus]="true">
+                    <igx-select [(ngModel)]="cell.editValue" width="100%" [igxFocus]="true">
                         <igx-select-item *ngFor="let race of races" [value]="race">
                             {{ race }}</igx-select-item>
                     </igx-select>
                 </ng-template>
             </igx-column>
-            <igx-column field="class" header="Class" [editable]="true" [filterable]="false" width="22%">
+            <igx-column field="class" header="Class" [editable]="true" [filterable]="false" width="20%">
                 <ng-template igxCellEditor let-cell="cell" let-value>
-                    <igx-select [(ngModel)]="cell.editValue" width="100px" [igxFocus]="true">
+                    <igx-select [(ngModel)]="cell.editValue" width="100%" [igxFocus]="true">
                         <igx-select-item *ngFor="let class of classes" [value]="class">
                             {{ class }}</igx-select-item>
                     </igx-select>
                 </ng-template>
             </igx-column>
-            <igx-column field="age" header="Age" [dataType]="'number'" [editable]="true" width="12%">
+            <igx-column field="age" header="Age" [dataType]="'number'" [editable]="true" width="13%">
             </igx-column>
             <igx-column field="alignment" header="Alignment" [editable]="true" [filterable]="false" width="22%">
                 <ng-template igxCellEditor let-cell="cell" let-value>
-                    <igx-select [(ngModel)]="cell.editValue" width="100px" [igxFocus]="true">
+                    <igx-select [(ngModel)]="cell.editValue" width="100%" [igxFocus]="true" [type]="'line'">
                         <igx-select-item *ngFor="let alignment of alignments" [value]="alignment">
                             {{ alignment }}</igx-select-item>
                     </igx-select>
