+++ src/app/grid/grids-routing.module.ts
@@ -111,7 +111,7 @@ import {
 } from "./multi-column-headers-styling/multi-column-headers-styling.component";
 import { GridSortingIndicatorsComponent } from './grid-sorting-indicators/grid-sorting-indicators.component';
 import { GridExportVisualizationComponent } from "./grid-export-visualization/grid-export-visualization.component";
-import {GridSelectComponent} from './grid-select/grid-editing-sample.component'
+import {GridSelectComponent} from './grid-select/grid-select-sample.component'
 // tslint:enable:max-line-length
 
 export const gridsRoutes: Routes = [
