+++ src/app/grid/grids.module.ts
@@ -120,7 +120,7 @@ import { GridMultiColumnHeadersStylingComponent } from "./multi-column-headers-s
 import { GridMultiColumnHeadersComponent } from "./multi-column-headers/multi-column-headers";
 import { GridSortingIndicatorsComponent } from './grid-sorting-indicators/grid-sorting-indicators.component';
 import { GridExportVisualizationComponent } from './grid-export-visualization/grid-export-visualization.component';
-import {GridSelectComponent} from './grid-select/grid-editing-sample.component';
+import {GridSelectComponent} from './grid-select/grid-select-sample.component';
 
 @NgModule({
     declarations: [
